src/bank-asset-agent/requirements.txt
grpcio
grpcio-tools
protobuf
PyJWT
cryptography
google-cloud-bigquery
requests
psycopg2-binary
sqlalchemy